[wip] scan: optimize cache lookup
Instead of updating the "last hit" index only on match, we should be
able to set it to the index that *would* have contained the element if
it had been there.

As we only search sorted lists in sorted order, we only ever need to
look past this element in future queries.