scan: make tests pass on windows
Ensure that all tests than can do pass on Windows. A few tests related
to symlinking are skipped because of technical limitations that aren't
trivial to overcome.

As Windows is more picky about not allowing open files to be deleted,
tests now properly close temporary files.

Added CI target for Windows as part of this as well.