[wip] scan: add '--out' flag for setting output file
The purpose of this flag is to enable the cache file to be used as
output file as well. Redirecting output to the file doesn't work because
that truncates the file before it's been read.

The current implementation adds the `--out` flag. The file is only
opened (and truncated) after the command has finished. This alleviates
the problem described above, but also means that any problem opening
the file will surface after having wasted a lot of computation. A simple
solution to this would be to just print on stdout, but there may be too
much output to fit in the buffer. A temporary file is therefore created
instead.

A message is logged if `--out` and `--cache` are the same. It's not
entirely clear if this is the best CLI to support the use case:
It feels like it shouldn't be allowed to have `--out` and `--cache`
point to the same file. But on the other hand, this is the exact use
case for the feature.