CI: Test against recent Go toolchain versions (#3)
Include versions 1.17 through 1.19 such that all "supported" versions are included.

Version 1.17 had some change that caused the 'path' value from 'filepath.Walk' to use backslashes instead of forward slashes on Windows. This broke a test that checks for the precense of a log entry which includes the path. The fix was just a matter of cleaning the path in the test though.