Minor refactorings
Mostly very small improvements picked from the implementation branch of the "match" command.

The only thing that's (loosely) relevant to this branch is the moving of a few functions to the file 'shared.go': The functions for the different commands will probably live in separate packages in the future, and this one will be shared between them.