scan: Fix return when file is not found
If a file is not found during scan (which is a situation that we cannot really provoke in provoke in tests, but does happen in practice) we only logged that it couldn't be found - we didn't prevent the error from getting returned and crashing the scan.

This adds a `nil` return to allow the scan to continue after such an error, and restructures the checking to call the error rewriting function earlier. Did a little naming improvements in these functions while at it.