scan: Replace testdata with generator system
Generating the exact data needed by the test is much more flexible than trying to define a single set of data that fits everywhere. It also eliminates the chance of tests interfering as they make slight modifications to the data (such as adding symlinks).

The data is expressed as a simple "spec" like
```go
root := dir{
    "a":   file{c: "x\n"},
    "c":   file{c: "y\n", inaccessible: true},
    "b/d": file{c: "x\n"},
    "e": dirExt{
        skipped: true,
        dir: dir{
            "f": dir{
                "a": file{c: "z\n", skipped: true},
                "g": file{},
            },
        },
    },
}
```

From this we automatically derive both data to be written on disk and also the expected scan result (I'm quite pleased with this, honestly).

While it may take away a little clarity from the test that we don't explicitly provide the expected result, it makes it trivial to add and iterate on tests (and also much smaller) - it might even enable property-based testing?! Also, some tests weren't exactly easy to understand before. It is kind of magic though, so to gain confidence that it's correct, we of course need to test the test system itself. This is to be done in a followup.

The old `testdata` directory has been deleted and all (scan) tests migrated to the new system.

Additional changes:
- scan: Add validation to `Dir` and `File` that names aren't empty. This will trigger if test data uses empty paths or paths ending with '/' (TODO: verify).
- hash: Extract function `New` for abstracting away the particular hash function that we use. Hopefully this means that we only have to change this single line if we ever want to use a different hash function. Also added `Bytes` for hashing a byte slice (used by test data generator).
- testutil: Consolidate `MakeFileInaccessible` and `MakeDirInaccessible` into single function `MakeInaccessible` as they were basically identical. I suspect they were separate only because there was also `MakeDirAccessible` which is also removed after being made obsolete by `T.TempDir()` - a fact which btw. is verified by new tests in `testutil/file_test.go`.