Generate testdata instead of checking it in (#12)
Generating the exact data needed by the test is much more flexible than trying to define a single/few file hierarchies that include all cases that we want to test. It also prevents interference between tests, as some might make slight modifications to the data (such as adding symlinks or making nodes inaccessible). So far it's worked fine because every test has cleaned up after itself properly, but it's not great practice (at the very least it eliminates any hope of running tests concurrently).

At least as important, this change will allow us to (easily) control the test files' timestamps which is to be included in the file matching (during scan) as required by #7.

The test data is expressed as a simple "spec" like
```go
root := dir{
    "a":   file{c: "x\n"},
    "c":   file{c: "y\n", makeInaccessible: true},
    "b/d": file{c: "x\n", hashFromCache: 42},
    "e": dirExt{
        skipped: true,
        dir: dir{
            "f": dir{
                "a": file{c: "z\n", skipped: true},
                "g": file{}, // empty file
                "h": dirExt{}, // empty directory
            },
        },
    },
}
```

From this we automatically derive both data to be written on disk (`root.writeTestdata(t, path)`) and also the expected scan result (`root.simulateScan(name)`, where `name` is the name of `root` in it's parent directory).

While it may take away some clarity from the test that it doesn't explicitly include the expected result, it makes it trivial to add tests and iterate on existing ones. They're also smaller and simpler, making it much easier to understand what is (and isn't!) being tested - this part has improved greatly. T It's even possible that the ability to derive expectations might enable property-based testing.

It is kind of magic though, so to gain confidence in correctness, we need to test the new test system also. This is to be done in a followup. A few other checks and improvements have been left as TODOs to be done as part of this.

The old `testdata` directory has been deleted and all (scan) tests migrated to the new system. For now they mostly just recreate the old data. Customizing the data for the particular test is another good case for followup improvements.

Additional changes:
- scan: Added validation to `Dir` and `File` that names aren't empty. This should trigger if test data uses empty paths or paths ending with '/' (TODO: verify).
- hash: Extracted function `New` for abstracting away the particular hash function that we use. Hopefully this means that we only have to change this single line if we ever want to use a different hash function. Also added `Bytes` for hashing a byte slice (used by test data generator).
- testutil: Consolidated `MakeFileInaccessible` and `MakeDirInaccessible` (and `MakeDirAccessible`) into single function `MakeInaccessible` as they were basically identical. It returns a function for reverting the change. The wrapper function `MakeInaccessibleT` takes care of error checking and cleanup registration for the purpose of tests. The ability for `T.TempDir` to clean up affected files and directories is verified by new tests in `testutil/file_test.go`.