scan: Follow symlinks in tests and verify/improve logging
Extracted helper function `tempDir` for ensuring that symlinks in temp dirs are always evaluated. This enables us to better verify logging without further hassles in all tests where it makes sense. As part of this, the test `Test__skipped_files_are_logged` was folded into `Test__skip_nonempty_files`.

Also, a warning is now logged if the hash of a file is cached with value 0.