scan: Output: Add modification time of files
The modification time may be optionally specified for test files. If omitted, the timstamp of generated files is not changed. To match this, the timestamp of the expected scan result is loaded from the actual file on disk. This is not super elegant, and requires us to extend the `node` interface to pass the complete parent path down the stack (by adding param `parentPath` to `simulateScanFromParent`) and also handle the error if the file cannot be "stat"ed. For now, we simply panic to avoid having to return an error that wouldn't be very useful anyway.

It also introduces a dependency on calling `writeTestdata` first (which we already do) unless the data already exists.

An alternative to all this is doing custom comparison that simply ignores `ts` when it's "expected" to be zero. In addition to avoiding the annoyances above, it might make assertion failures easier to understand to get just the first failure in a depth-first search compared to the complete diff. This wouldn't catch if scan actually returns a zero timestamp, but we can just test that separately.