cmd: Fix 'scan' tests
Some tests in this package run 'scan' on testdata files which are checked into the repo. These tests broke with the introduction of modification time into the result file in the previous commit. The fix is to read the time from the actual file. It's deemed fine to add this variability into the tests as we still test against the 'os' package, with is trusted.

This work led to the following loosely related changes being made in `checkCache` (and reflected in other tests of the same package):
- Verify that directory names are not empty.
- Don't check ordering of empty files as these aren't used in caching.
- Updated comments.

Also, the test `Test__scan_testdata_with_trailing_slash` was merged into `Test__scan_testdata` and more cases were added.