scan: test: Compare result using custom assertion
When defining files for a "scan" test case (using the `node` interface), we can either specify a modification time of the file or leave it as zero, in which case it defaults to the time of running the test.

In order to be able to compare scan result (`Dir`) using `assert.Equal`, we previously patched the expectation with the actual mod time of the file. This complicates the `node` interface a bit and generally doesn't taste right. With this commit we replace this solution with custom methods on `Dir` and `File` (it turns out that you can add test-only methods to structs!). The methods behave almost like `assert.Equal` with the exception that mod times are assumed equal if the expected value is zero. This idea was also described in 3e683de, where the previous solution was originally implemented.

It's a little unfortunate that there isn't just some way for us to override equality for a type within the confines of `assert.Equal`...