scan: test: Reject dir entries with overlapping paths
`dir` is defined as a map which has unique key`dir` is defined as a map, which has unique keys. But these keys can contain '/' to denote a nested path, introducing the possibility that the paths overlap. While this could be allowed, it introduces ambiguity in use, is likely a mistake, and requires more logic than it's worth (like additional sorting) in `simulateScan` to handle all cases correctly. For `writeTestdata` it works automatically and would add complexity to make it fail. So we add a check to reject in the former and allow it in the latter (even duplicate files).