scan: test: Let writeTestdata reject duplicate files
In 5c899b1 we decided that duplicate/overlapping paths were disallowed, but wouldn't bother checking for it in `writeTestdata`. However, we did add a test `Test__node_with_overlapping_files/writeTestdata` that is works. However, this test was flaky because `writeTestdata` doesn't iterate `dir` entries in any particular order (as that's generally unnecessary for valid input).

I'm unsure if testing that something invalid works is even good practice, but now we test that it fails as expected instead (which is definitely good).

It's really strange that this wasn't caught by the CI in #20, but for some reason the tests didn't get included in the run for the final commit (which is where this was introduced).