scan: Make more characters invalid to use in skip expressions (#22)
We reject characters that have a common special meaning in relation to
certain features, while being very uncommon (or even invalid) in file
names:

- OS-specific path separator (in addition to `/` which is already
invalid): Ensure that `\` isn't used on Windows. The purpose of
rejecting these is to avoid giving the user the impression that nested
paths are supported.

- `*` and `?`: Their usage is an indicator that the user intended to use
regex/glob expressions. It's likely that we're going to support such
expressions in the future.