Update .gitignore and migrate remaining testdata
Removed `.gitattributes`; it was added to preserve the contents of testdata before #12 made it generated instead.

We still scan the checked-in files in `cmd/dupe-nukem/testdata`, so it should probably contain a local `.gitattributes`. However, before adding that, we want to see the CI fail on Windows to verify expected behavior.