[WIP] Add type for safe path lookup in tests using `node`
Make `writeTestdata` (of `node`) return a value of new type `testdata` that knows the path on which the testfile was written to disk. Is has a method `find` for navigating the tree of files. This allows us to construct paths in tests "safely" instead of using `filepath.Join`, which is basically string concatenation.

While this works, it only knows about the literal keys used in `dir`; if the keys have nested paths, we have to navigate that atomically (i.e. if it has path `x/y`, we cannot "find" just `x`). One the one hand, this shortcoming makes it insufficiently general. On the other, it feels like we're already on the limit of the justifiable amount of complexity for the problem at hand.

This branch is therefore not intended to be merged, but only left here in case other circumstances makes the proper solution worth completing later on.