scan: Ignore skipping of root (#26)
Instead of failing with an error when the skip list includes the root,
we now just log that it isn't going to be skipped.

If the root is a symlink, then skipping applies to its target. The
symlink resolution is logged separately, so this should not be a source
of surprise.