scan: test: Restore logger on cleanup
Currently, once we've started capturing logs, the original state of the logger is never restored. This is not a big problem, as we always configure it in the tests where we check logs, but it seems appropriate to restore modified global state (and in turn not keep objects from previous tests alive).

Note that we can eliminate logs from the test output by setting the output to a discarding writer in an `init` function as the cleanup function is going to restore this.

We now also explicitly check that we don't capture the logs in parallel tests as that won't work as long as the logger is global.

Finally, the function name was changed from `CollectLogs` to `CaptureLogs` to reflect that we don't just collect the logs that are being emitted, but consume them entirely.