CI: Make linter failure fail the workflow
It turns out that by default, `golint` exits "successfully", even when violations were reported. GitHub does pick up the output, but only displays it in annotations to the workflow summary (and PR diff) without failing the workflow. There's no automatic notification, making it easy to miss.

This commit adds a flag to make the linter exit with code 1 (thus failing the workflow) if any violations are reported.