Add version to scan result and emit abs root path
The result `Dir` is wrapped in a new type `Result` with the previous result in key `root` and an integer field `version` for the format version. The version starts at 1 to ensure that default decoded value of 0 is never treated as valid.

The root dir name now uses the full path with symlinks completely resolved. The root name of caches must thus match on the absolute, resolved path.

Resolves #19.