scan: Add version to result and save absolute path of root (#31)
The result `Dir` is wrapped in a new type `Result` with the previous
result in key `root` and an integer field `schema_version` for version
of the data model. The version starts at 1 to ensure that default
decoded value of 0 isn't mistaken for an explicit version.

The root dir name now uses the full path with symlinks completely
resolved. The root name of caches must thus match on the absolute,
resolved path.

Resolves #19 and #28.