scan: Log running time
This both exposes a useful metric to the user and signals that the scan did complete and make it easier to locate the cause of any subsequent error.

There's some problem for large scans that makes it crash (seemingly with OOM) during JSON serialization of the output. Adding a little instrumentation doesn't reveal the problem, but it might be that we have to manually stream the JSON instead of relying on the builtin functionality.