Initial work on web-based matcher
Implemented loading of example scan results produced by a new micro tool `gendata` that uses the newly extracted testdata setup to simulate before/after data to match against. This in turns converted the project to a "workspace" project, making it easy for one workspace to depend on another.

While the vite project is in plain JS, we add static type checking using JSDoc. Type errors are displayed as errors in the editor (which behind the scenes actually uses TypeScript - hence the addition of `tsconfig.json` to make it work correctly). The checking can also be run via `tsc` on the command line if at some point we want to add a CI step for this. At any rate, no transpilation is performed, meaning that types don't affect runtime at all.

The app currently does the following:
- The scan results are loaded directly from disk at the hardcoded path and interpreted as `ScanResult` (literally `scan.Result`) without any input validation.
- This object is translated into a `Target`, in which the files are indexed by hash and all objects are wrapped in `Target*` counterparts which include pointers to their parent. These objects also initialize the DOM elements into which they're going to be rendered. It isn't clear whether the `Scan*` objects should be retained as part of the targets, but we're keeping them for now.
- The resulting targets and the DOM nodes are dumped into the document...