Convert domain types to classes, validate JSON input, and start improving styling
Classes seem more convenient for this use case as they are `instanceof`'able and because it seems plausible that we're going to want the objects to have instance methods.

The input is now entirely type safe by marking types as `unknown` until the values have been validated.

The page is styled such that the "app" part is the only scrollable part, with header and footer being always visible.