Add rendering of dirs that contain any matching files
A nice improvement would be to grade the color by how deeply you have to go into the dir to find a match.

It's becoming clear that `findMatchesOf` (and other methods like it to follow) should operate on Dir/File objects, not their Dom variants. The Dom objects only have to come into play when it's time to call `mark` on them. That refactoring should be done next.