Add marking on files if they have no matches in either their own nor other targets
This is currently done by the controller in a new method `init`. It would probably be better to do a pass on the domain model after it's been built and annotate the objects with this kind of information.

We could then defer DOM initialization until afterwards and have the marks applied already at that point. On the other hand, leaving it to the controller would allow us to dynamically change settings like whether to consider matches from the target itself. So we should probably keep it to the above paragraph.