Only update matching when holding the alt key
The only minor quirk is that if you press the alt key while hovering outside the target container, then the matching doesn't clear. Two simple solutions to this that come to mind are (1) clear on alt keydown or (2) register the handler on the entire window. But it isn't worth worrying about for now.