Fixed quirk mentioned in previous commit
Yes, the one that wasn't worth addressing for now... But it also fixed another minor quirk that you would only "hover" an element after actively moving the curser over it from outside the element.

The solution is - as always - a little more complex than expected: The keypress handler doesn't have access to the currently hovered element. So we store that in a new controller state `deferredTarget` to be used when calling from the keypress handler. At some point we should refactor these handlers to reflect what they do rather than what they react on...